Put the plugin in a file called ZonesPlugin.js
